Cleanup encoding Startup message
Further cleanup of message encoding:

- Move Startup struct into PostgresFrontendMessageEncoder
- Move PSQLMessagePayloadEncodable into tests, since it isn't used in PostgresNIO anymore
- Only support the parameters that are actually used in encoding startup messages
