+++ Tests/PostgresNIOTests/New/Extensions/PSQLBackendMessageEncoder.swift
@@ -257,3 +257,7 @@ extension RowDescription: PSQLMessagePayloadEncodable {
         }
     }
 }
+
+protocol PSQLMessagePayloadEncodable {
+    func encode(into buffer: inout ByteBuffer)
+}
