+++ Tests/PostgresNIOTests/New/Messages/StartupTests.swift
@@ -4,56 +4,44 @@ import NIOCore
 
 class StartupTests: XCTestCase {
     
-    func testStartupMessage() {
+    func testStartupMessageWithDatabase() {
         var encoder = PostgresFrontendMessageEncoder(buffer: .init())
         var byteBuffer = ByteBuffer()
-        
-        let replicationValues: [PostgresFrontendMessage.Startup.Parameters.Replication] = [
-            .`true`,
-            .`false`,
-            .database
-        ]
-        
-        for replication in replicationValues {
-            let parameters = PostgresFrontendMessage.Startup.Parameters(
-                user: "test",
-                database: "abc123",
-                options: "some options",
-                replication: replication
-            )
-            
-            encoder.startup(parameters)
-            byteBuffer = encoder.flushBuffer()
 
-            let byteBufferLength = Int32(byteBuffer.readableBytes)
-            XCTAssertEqual(byteBufferLength, byteBuffer.readInteger())
-            XCTAssertEqual(PostgresFrontendMessage.Startup.versionThree, byteBuffer.readInteger())
-            XCTAssertEqual(byteBuffer.readNullTerminatedString(), "user")
-            XCTAssertEqual(byteBuffer.readNullTerminatedString(), "test")
-            XCTAssertEqual(byteBuffer.readNullTerminatedString(), "database")
-            XCTAssertEqual(byteBuffer.readNullTerminatedString(), "abc123")
-            XCTAssertEqual(byteBuffer.readNullTerminatedString(), "options")
-            XCTAssertEqual(byteBuffer.readNullTerminatedString(), "some options")
-            if replication != .false {
-                XCTAssertEqual(byteBuffer.readNullTerminatedString(), "replication")
-                XCTAssertEqual(byteBuffer.readNullTerminatedString(), replication.stringValue)
-            }
-            XCTAssertEqual(byteBuffer.readInteger(), UInt8(0))
-            
-            XCTAssertEqual(byteBuffer.readableBytes, 0)
-        }
+        let user = "test"
+        let database = "abc123"
+
+        encoder.startup(user: user, database: database)
+        byteBuffer = encoder.flushBuffer()
+
+        let byteBufferLength = Int32(byteBuffer.readableBytes)
+        XCTAssertEqual(byteBufferLength, byteBuffer.readInteger())
+        XCTAssertEqual(PostgresFrontendMessage.Startup.versionThree, byteBuffer.readInteger())
+        XCTAssertEqual(byteBuffer.readNullTerminatedString(), "user")
+        XCTAssertEqual(byteBuffer.readNullTerminatedString(), "test")
+        XCTAssertEqual(byteBuffer.readNullTerminatedString(), "database")
+        XCTAssertEqual(byteBuffer.readNullTerminatedString(), "abc123")
+        XCTAssertEqual(byteBuffer.readInteger(), UInt8(0))
+
+        XCTAssertEqual(byteBuffer.readableBytes, 0)
     }
-}
 
-extension PostgresFrontendMessage.Startup.Parameters.Replication {
-    var stringValue: String {
-        switch self {
-        case .true:
-            return "true"
-        case .false:
-            return "false"
-        case .database:
-            return "replication"
-        }
+    func testStartupMessageWithoutDatabase() {
+        var encoder = PostgresFrontendMessageEncoder(buffer: .init())
+        var byteBuffer = ByteBuffer()
+
+        let user = "test"
+
+        encoder.startup(user: user, database: nil)
+        byteBuffer = encoder.flushBuffer()
+
+        let byteBufferLength = Int32(byteBuffer.readableBytes)
+        XCTAssertEqual(byteBufferLength, byteBuffer.readInteger())
+        XCTAssertEqual(PostgresFrontendMessage.Startup.versionThree, byteBuffer.readInteger())
+        XCTAssertEqual(byteBuffer.readNullTerminatedString(), "user")
+        XCTAssertEqual(byteBuffer.readNullTerminatedString(), "test")
+        XCTAssertEqual(byteBuffer.readInteger(), UInt8(0))
+
+        XCTAssertEqual(byteBuffer.readableBytes, 0)
     }
 }
